Add more capabilities to the CMake build (#4163)
This commit includes changes that enable the following in the CMake build:
* Building protobuf from source.
* Building grpc from source.
* Building the TensorFlow distributed runtime.
* Building the TensorFlow Python bindings (excluding tf.contrib).

This commit also includes minor changes to remove the implicit
dependency between a CPU-only build of the runtime and stream executor,
and makes minor changes for compatibility with Protobuf 3.0.0 (regarding
compatibility between Protobuf's and TensorFlow's int64 representation).

+++ tensorflow/core/distributed_runtime/rpc/grpc_remote_worker.cc
@@ -132,8 +132,8 @@ class GrpcRemoteWorker : public WorkerInterface {
             // the RecvTensor response can not have been sent before
             // the RecvTensor request, and must have been sent before
             // it was received.
-            send_start_usec =
-                std::max(start_usec, response->metadata().send_start_micros());
+            send_start_usec = std::max(start_usec, static_cast<int64>(
+                response->metadata().send_start_micros()));
             send_start_usec = std::min(send_start_usec, end_usec - 1);
           }
           const string& key = request->rendezvous_key();
